refactor: clean up agent panel — rename rubi → montte-ai, drop dead json-render
- Rename `-rubi-panel/` → `-montte-ai/`; `agent-panel.tsx` → `panel.tsx`; `agent-chat-store.ts` → `chat-store.ts`
- Consolidate 4 tool-renderer files (advisor, lazy, skill, services) into single `tool-renderers.tsx`
- Inline `<img src="/mascot.svg">` at the 2 use sites; delete `agent-mascot-icon.tsx`
- Drop unused `@json-render/{core,react,shadcn}` packages and frontend registry/catalog
- Drop `modules/agents/src/json-render.ts` and `RENDERING_PRIMER` (no tool ever returned `ui` fields — was misleading the model)
- Clean up dead exports in `-context-panel/` (toggleContextPanel, registerTab, dynamicTabs, pageViewSwitch, god hook); fix no-dep `useEffect` anti-pattern via ref + `useIsomorphicLayoutEffect`
