--- apps/web/src/routes/_authenticated/$slug/$teamSlug/-rubi-panel/agent-mascot-icon.tsx
@@ -1,17 +1,0 @@
-import type { ComponentProps } from "react";
-
-export function AgentMascotIcon({
-   className,
-   ...props
-}: ComponentProps<"img">) {
-   return (
-      <img
-         alt=""
-         aria-hidden="true"
-         className={className}
-         draggable={false}
-         src="/mascot.svg"
-         {...props}
-      />
-   );
-}
