--- apps/web/src/routes/_authenticated/$slug/$teamSlug/-rubi-panel/json-render/registry.tsx
@@ -1,27 +1,0 @@
-import { defineRegistry } from "@json-render/react";
-import { shadcnComponents } from "@json-render/shadcn";
-import { agentCatalog } from "./catalog";
-
-export const { registry: agentRegistry } = defineRegistry(agentCatalog, {
-   components: {
-      Card: shadcnComponents.Card,
-      Stack: shadcnComponents.Stack,
-      Grid: shadcnComponents.Grid,
-      Separator: shadcnComponents.Separator,
-      Heading: shadcnComponents.Heading,
-      Text: shadcnComponents.Text,
-      Badge: shadcnComponents.Badge,
-      Alert: shadcnComponents.Alert,
-      Table: shadcnComponents.Table,
-      Accordion: shadcnComponents.Accordion,
-      Collapsible: shadcnComponents.Collapsible,
-      Tabs: shadcnComponents.Tabs,
-      Progress: shadcnComponents.Progress,
-      Skeleton: shadcnComponents.Skeleton,
-      Spinner: shadcnComponents.Spinner,
-      Avatar: shadcnComponents.Avatar,
-      Image: shadcnComponents.Image,
-      Link: shadcnComponents.Link,
-      Tooltip: shadcnComponents.Tooltip,
-   },
-});
